refactor: make soobook entity which contains column [id, createdAt, updatedAt, updatedIp]
